Moved Author page to Author folder and created author input page
