--- CEN4020_Website/Pages/Author.cshtml
@@ -1,11 +1,0 @@
-﻿@page
-@model CEN4020_Website.Pages.AuthorModel
-@{
-    ViewData["Title"] = "Submission";
-}
-
-<div>
-    <br>
-    <h2>Paper Submission</h2>
-    
-</div>
