+++ CEN4020_Website/Pages/Author/Index.cshtml
@@ -1,0 +1,16 @@
+﻿@page
+@model CEN4020_Website.Pages.Author.IndexModel
+@{
+}
+
+<br>
+<h2>New Paper Submission</h2>
+
+<form method="POST">
+    <div class="row mb-3">
+        <label class="col-sm-3 col-form-label">Title</label>
+        <div class="col-sm-6">
+            <input type="text" class="form-control" name="title" value="">
+        </div>
+    </div>
+</form>
